!BETA! gummiboot support (nneds review). EPS set at /boot
I decided to add the beginnings of what could be Gummiboot support to the script.  I also changed the EFI mount back to /boot in order to a)make Gummiboot install easier b) address lack of adoption of /boot/efi standard for ESP.

Please note that this is a highly dangerous fork of helmuth's original Arch Ultimate Install meant to solely address the lack of Gummiboot support.  I am not going to become an active developer of the aui project.  The implementation is in testing, and will hopefully make an appearance in the main aui project at some point.

Please comment on any issues you may have with the release.

Per Laborem ad Astra,
JbstormburstADV